Add files via upload
Version 3.2.2

Major Changes by JSantner:

PipeSolver calculates a huge pressure drop, such that the pressure is negative at the exit of the tubing. So, you have an unphysically high mass flow rate, that cannot exist with a tube of this length and diameter.

To deal with this, I think that your method of solving for the mass flow shouldn't be used in filling mode. I'm trying a different method based on The Bernoulli equation here.